requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
phonepe==1.0.1
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
